Added player behaviour
Made the player able to move, and made them able to spawn and join/host a game. Also started making some seat functions, so that I could easily implement the ball later. Spawning object's gets handled with RPC and command's (To and from the server). I chose this, because with this method every player sees the same things.