Ball destruction
When de ball hits either of the walls behind the player, the ball gets destroyed. After the destruction, the server gets a command to update the scores of the clients. After recieving the scores, the server sends a RPC back to the clients. This way the clients both have the same score on their screen.