Fixed sync problems
Had some syncing problems with the ball. After brainstorming for a bit, I managed to fix it, by sending commands and RPC's to the server and clients. So that everything has the same values. Before I did this, the ball sometimes took other angles if players had high latency/ping. This was caused by the ball not syncing properly. So I stopped using the UNET component, and started syncing with commands.